+++ esp32/snapclient/include/boot_mode_selector.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "snapclient_config.h"
+
+app_config::OperatingMode detectOperatingMode();